+++ utils.c
@@ -17,7 +17,7 @@ inline size_pt inchToPt(const size_in size) {
 }
 
 inline size_in milToInch(const size_mil size) {
-    return (float) size / 1000;
+    return (double) size / 1000;
 }
 
 inline bool strEqual(const char *s1, const char *s2) {
